requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai==1.10.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
